+++ server.js
@@ -1,8 +1,8 @@
 const express = require('express');
-const puppeteer = require('puppeteer');
 const bodyParser = require('body-parser');
 const { sequelize } = require('./models');
 const testRoutes = require('./routes/testRoutes');
+const proxyRoutes = require('./routes/proxyRoutes');
 const swaggerUi = require('swagger-ui-express');
 const swaggerJsdoc = require('swagger-jsdoc');
 const cors = require('cors')
@@ -39,82 +39,10 @@ app.use(bodyParser.json());
 
 app.use('/api', testRoutes);
 
+app.use('/proxy', proxyRoutes)
+
 app.get('/', (req, res) => {
   res.send('Nothing to see here. Please refer to /api-docs or GitHub repo (https://github.com/ESCORIAL-SAIC/mbti-api) for documentation.');
-});
-
-app.get('/proxy/test/:id', async (req, res) => {
-  const id = req.params.id;
-  const url = `https://devil-ai.translate.goog/api-personality-test/${id}?_x_tr_sl=en&_x_tr_tl=es&_x_tr_hl=es&_x_tr_pto=wapp`;
-
-  try {
-    const browser = await puppeteer.launch({
-      headless: true,
-      args: ['--no-sandbox', '--disable-setuid-sandbox']
-    });
-    const page = await browser.newPage();
-    await page.goto(url, { waitUntil: 'networkidle0' });
-
-    await page.evaluate(() => {
-      const style = document.createElement('style');
-      style.innerHTML = `
-        iframe#gt-nvframe,
-        iframe[src*="translate.google.com"] {
-          display: none !important;
-        }
-        body {
-          margin-top: 0 !important;
-        }
-      `;
-      document.head.appendChild(style);
-    });
-
-    const content = await page.content();
-    await browser.close();
-
-    res.setHeader('Content-Type', 'text/html');
-    res.send(content);
-  } catch (error) {
-    console.error('Error:', error);
-    res.status(500).send('Error al cargar la página traducida con Puppeteer');
-  }
-});
-
-app.get('/proxy/result/:id', async (req, res) => {
-  const id = req.params.id;
-  const url = `https://devil-ai.translate.goog/r/${id}?_x_tr_sl=en&_x_tr_tl=es&_x_tr_hl=es&_x_tr_pto=wapp`;
-
-  try {
-    const browser = await puppeteer.launch({
-      headless: true,
-      args: ['--no-sandbox', '--disable-setuid-sandbox']
-    });
-    const page = await browser.newPage();
-    await page.goto(url, { waitUntil: 'networkidle0' });
-
-    await page.evaluate(() => {
-      const style = document.createElement('style');
-      style.innerHTML = `
-        iframe#gt-nvframe,
-        iframe[src*="translate.google.com"] {
-          display: none !important;
-        }
-        body {
-          margin-top: 0 !important;
-        }
-      `;
-      document.head.appendChild(style);
-    });
-
-    const content = await page.content();
-    await browser.close();
-
-    res.setHeader('Content-Type', 'text/html');
-    res.send(content);
-  } catch (error) {
-    console.error('Error:', error);
-    res.status(500).send('Error al cargar la página traducida con Puppeteer');
-  }
 });
 
 sequelize.sync();
